% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_codes.R
\name{all_codes}
\alias{all_codes}
\title{All SARB available codes}
\usage{
all_codes(dateformat = FALSE, token = NULL)
}
\value{
list
}
\description{
Returns all available codes in the DB
}
\examples{
\dontrun{
all_codes()
}
}
