% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency_description.R
\docType{data}
\name{frequency_description}
\alias{frequency_description}
\title{South African Reserve Bank frequency code metadata}
\format{
A data.frame with columns Description, Frequency
}
\source{
South African Reserve Bank
}
\description{
metadata of frequencies as provided per the SARB \url{https://www.resbank.co.za/Pages/default.aspx}
}
