% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator_codes.R
\docType{data}
\name{indicator_codes}
\alias{indicator_codes}
\title{South African Reserve Bank indicator metadata}
\format{A data.frame with columns time_series_number, description_and_version, series_code, version_description, frequency, unit_of_measure, full_ind_number}
\source{
South African Reserve Bank
}
\description{
metadata of all indicators as provided per the SARB \url{https://www.resbank.co.za/Pages/default.aspx}
}
