% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release_info.R
\name{release_info}
\alias{release_info}
\title{Release Information}
\usage{
release_info(dateformat = FALSE)
}
\arguments{
\item{dateformat}{return latest release information as date object}
}
\value{
list
}
\description{
Returns the latest release date of the data
}
\examples{
\dontrun{
release_info(dateformat = FALSE)
release_info(dateformat = TRUE)
}
}
