% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarb_code.R
\name{sarb_code}
\alias{sarb_code}
\title{SARB code}
\usage{
sarb_code(code, token = NULL)
}
\value{
list
}
\description{
Request SARB data using the designated code with format KBPXXXX
}
\examples{

sarb_code(code = "KBP1434D", token = "your token here")

options(sarb.token = "your token here")
sarb_code(code = "KBP1434D")

Sys.setenv("sarb.token" = "your token here")
sarb_code(code = "KBP1434D")
}
