% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token_request.R
\name{token_request}
\alias{token_request}
\title{Token Request}
\usage{
token_request()
}
\value{
list
}
\description{
Request a token from the server to use the api service
}
\examples{

token <- token_request()
print(token)

}
