% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarb_code.R
\name{sarb_code}
\alias{sarb_code}
\title{SARB code}
\usage{
sarb_code(code, token = NULL)
}
\value{
list
}
\description{
Request SARB data using the designated code with format KBPXXXX. This function is the worker function of the `sarbR` package and is used to interact with an API linked to the main dataabse
}
\examples{
sarb_code(code = "KBP6045J", token = "9eadf0a7a345a0cd286d81e74c555715")

Sys.setenv("sarb.token" = "9eadf0a7a345a0cd286d81e74c555715")
options(sarb.token = "9eadf0a7a345a0cd286d81e74c555715")

sarb_code(code = "KBP6045J")

}
