% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarb_code.R
\name{sarb_code}
\alias{sarb_code}
\title{SARB code}
\usage{
sarb_code(code, token = NULL)
}
\value{
list
}
\description{
Request SARB data using the designated code with format KBPXXXX. This function is the worker function of the `sarbR` package and is used to interact with an API linked to the main dataabse
}
\examples{
sarb_code(code = "KBP1434D", token = "f84d879ce6537dbacd93ac3dc073e364")

Sys.setenv("sarb.token" = "f84d879ce6537dbacd93ac3dc073e364")
options(sarb.token = "f84d879ce6537dbacd93ac3dc073e364")

sarb_code(code = "KBP1434D")

sarb_code(code = "KBP1434D")

}
