% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_indicator.R
\name{search_indicator}
\alias{search_indicator}
\title{Seach Indicator}
\usage{
search_indicator(description, time_series_code)
}
\arguments{
\item{description}{description of indicator to look up}

\item{code}{code of indicator to look up}
}
\value{
data.frame
}
\description{
A function that searches all available indicators based on description input or code. If both are given then parameter code will take preference
}
\examples{

search_indicator(description = "GDP")
search_indicator(time_series = "KBP1000")
}
